requirements.txt
PyYAML~=6.0
python-telegram-bot~=20.1
gTTS~=2.3.1
python-forecastio~=1.4.0
geopy~=2.3.0
